requirements.txt
openai
simpleaudio
openai_stt
typing_extensions
huggingface-hub
playsound3
PyGithub